added routing and changed file locations
