+++ backend/package.json
@@ -11,8 +11,7 @@
     "express": "^4.19.2",
     "express-session": "^1.18.0",
     "firebase": "^10.12.3",
-    "fs": "^0.0.1-security",
-    "react-router-dom": "^6.26.1"
+    "fs": "^0.0.1-security"
   },
   "devDependencies": {
     "@types/express": "^4.17.21",
