--- frontend/src/pages/HomePage.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-
-
-function HomePage() {
-
-
-  return (
-    <>
-        Welcome Home!
-    </>
-  );
-}
-
-export default HomePage;
