+++ frontend/src/pages/HomePage/HomePage.scss
