+++ frontend/src/pages/HomePage/HomePage.tsx
@@ -1,0 +1,22 @@
+// @ts-ignore
+
+import PannellumBackground from "../../components/PannellumBackground/PannellumBackground";
+import "./HomePage.scss";
+
+function HomePage() {
+
+  return (
+    <div
+      style={{
+        display: "flex",
+        alignItems: "center",
+        height: "100%",
+        justifyContent:"center"
+      }}
+    >
+      HOME PAGE IN CONSTRUCTION!
+    </div>
+  );
+}
+
+export default HomePage;
