--- frontend/src/pages/PlayPage.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function PlayPage() {
-  return (
-    <>
-        Playing the game!
-    </>
-  );
-}
-
-export default PlayPage;
