+++ frontend/src/pages/PlayPage/PlayPage.tsx
@@ -1,0 +1,17 @@
+
+function PlayPage() {
+  return (
+    <div
+      style={{
+        display: "flex",
+        alignItems: "center",
+        height: "100%",
+        justifyContent:"center"
+      }}
+    >
+     PLAY PAGE IN CONSTRUCTION!
+    </div>
+  );
+}
+
+export default PlayPage;
