--- frontend/src/pages/ProfilePage.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-
-function ProfilePage() {
-  return (
-    <>
-        Here is all the information about you!
-    </>
-  );
-}
-
-export default ProfilePage;
