+++ frontend/src/pages/RegisterPage/RegisterPage.tsx
@@ -1,0 +1,9 @@
+import Register from '../../components/Forms/Register/Register';
+
+function RegisterPage() {
+  return (
+    <Register/>
+  );
+}
+
+export default RegisterPage;
